Fix remaining time in nanosleep
Add the fix of commit 96f68ab in unikraft to newlib:
Calculate the remaining time to sleep and update
the rem parameter if it is given.
If the remaining time is larger than 0, it means that
the thread was waken up explicitly and nanosleep returns -1
to indicate that. Otherwise nanosleep returns 0.

Signed-off-by: Dafna Hirschfeld <dafna3@gmail.com>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>