Adapt and use musl ioctl.h headers
The lwip library has its own definition for SIOCATMARK which we cannot
overwrite, so we avoid using our own in that case. Also, the same
definition for winsize structure is already provided by the musl
headers, so we stick to those.

This patch also adds licenses for the musl ioctl.h headers.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>