syslog.c: Import from osv
Copied as is from the official osv git mirror
https://github.com/cloudius-systems/osv.git, commit ee7a2cd4.

Original file locations:
* libc/syslog.c

We had to decide between importing from musl or osv. We decided to pick the osv
variant because it's simpler: it outputs to stdio. The musl implementation uses
UNIX sockets, which we currently don't have in Unikraft. Moreover, the osv
implementation was originally imported from musl and adapted. This is also the
reason why we put the file under the musl-imported/ subdirectory.

In the future, a better syslog implementation may replace this one and it might
even become an internal library with fully fledged functionality, just like
classic OSes.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>