musl-imported: Add original copyright
Given that there are and will be lots of files imported from musl, we add the
original copyright so that adding the licenses on the imported files will not be
necessary unless there will be a reasonable number of changes.

Copied as is from the official musl git mirror
git://git.musl-libc.org/musl, commit 6fcb440d.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Charalampos Mainas <charalampos.mainas@neclab.eu>