Revisit process.c functions
This patch revisits the functions in process.c, redistributes some of them and
implements the others. The remaining functions may very well be moved into an
internal "process" related library, but we will do this at another time. It's
worth saying that we set the PID of the unikernel to 1, saving the PID 0 for its
"parent", an invisibile process that is also not shown on Linux.

Redistributed functions:
- kill() and killpg() are moved to signal.c
- times() is moved to time.c

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>