Use our own <time.h> and <sys/time.h> headers
The uktime internal library provides the headers with the time related
definitions that we need. Given that newlib does not provide guard macros for
most of its definitions (and consequently we get errors because of duplicated
definitions), we decided to skip using its time headers.

In these glue headers we add only those definitions and inclusions that are
needed by newlib. For example, newlib's <sys/time.h> includes <sys/reent.h>, so
we will do the same here. In <time.h> we added the time zone/locale definitions
needed by newlib.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>