Config.uk: remove imply of shim layer library
Currently, Config.uk sets imply LIBSYSCALL_SHIM, which for all intents
and purposes forces all images built with newlib to include the
syscall shim even when it's not needed. While "imply" still allows
users to manually deselect the option, one has to know about this
possiblity, and kraft-driven builds will not do this by default.

For applications that *do* need the shim library, this should be set
explicitly in their Config.uk files, or in their respective kraft.yaml
file.

Signed-off-by: Felipe Huici <felipe.huici@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>