Define tzname
Currently, tzname is defined as an extern, which results in a
undefined reference linker error when attempting to use
it. This patch maps tzname to _tzanme, which *is*
defined. newlib does this mapping in OS-specific files that we
don't compile in, so this is a way to mimic that behavior.

Signed-off-by: Felipe Huici <felipe.huici@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>