include/fcntl.h: Replace O_NONBLOCK with new one on arm
in newlib _default_fcntl.h, it defines O_NONBLOCK as follows:
\#define _FNONBLOCK      0x4000  /* non blocking I/O (POSIX style) */
But it is different from linux definition(04000=0x800):
\#define O_NONBLOCK	00004000

This difference confused virtio-9p open() syscall. If you open a
file with O_NONBLOCK(is 0x4000 in newlib) and sys_open will regard
it as O_DIRECTORY. At last it returns a ENOTDIR error.

Signed-off-by: Jia He <justin.he@arm.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>