Correct license headers introduced by University Politehnica of Bucharest
This commit removes the additional clause
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY"
from BSD license headers that got introduced by mistake
with the following commits:
 949e761 - Use our own <time.h> and <sys/time.h> headers (Costin Lupu)
 b2192bd - Add device related stubs (Vlad-Andrei BĂDOIU (78692))
 2c79cd5 - Introduce <sys/unistd.h> for including both newlib and Unikraft definitions (Costin Lupu)
 04f03ec - Add WCOREDUMP and WIFCONTINUED (Costin Lupu)
 d901960 - Add sys/signal.h header (Mihai Pogonaru)
 9287428 - Add mntent.c stubs (Costin Lupu)
 20ab046 - Introduce <uk/_types.h> header for basic types defined by Unikraft (Costin Lupu)

On the one hand, this additional clause is redundant because
the BSD license already states that it must remain and the
copyright notice must be kept.
On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD 3-clause.

The original author(s) or an authorized representative from
the author's affiliation consents to the change with a
`Reviewed-by` tag to this commit.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #8