Makefile.uk: Use -ffreestanding as internal build option
As a standard C library (libc), newlib is to be built in a freestanding
environment (i.e. not relying on a libc). For that, build internal
newlib source code files with the `-ffreestanding` option. Otherwise, a
hosted environment will be used; in a hosted environment, certain
optimizations may be enabled by the compiler, and they may cause issues
(such as an infinite set of calls).

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>